Remove explicit reference to System.Net.Http
The assembly is only used in CloudMetadataService for .NET Standard
builds but System.Net.Http is built-in in .NET Standard and doesn't
need an explicit reference.

Solution mixing nuget and built-in references to System.Net.Http leads
to runtime issues that we are trying to solve by removing the nuget
references.

https://github.com/dotnet/runtime/issues/26131#issuecomment-388199236
> In most cases, we don't advise people use the separate System.Net.Http
> NuGet package anymore