docs: mention ScyllaDB alongside Cassandra
As the driver is forked from DataStax, there are mentions of Cassandra
and DataStax Enterprise in the documentation. This commit replaces
mentions of DSE, which we don't support, with ScyllaDB.

Also, it clarifies that ScyllaDB uses protocol v4 with custom
extensions, and that the optimisation related to caching prepared
statements metadata is in progress.